% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_contemporaneous_correlations.r
\name{contemporaneous_correlations_plot}
\alias{contemporaneous_correlations_plot}
\title{Plot the contemporaneous correlations summary}
\usage{
contemporaneous_correlations_plot(av_state)
}
\arguments{
\item{av_state}{an object of class \code{av_state} that was the result of a call to \code{\link{var_main}}}
}
\description{
This function plots a summary of the contemporaneous correlations found in the valid models found. It is called as part of \code{\link{var_summary}}.
}
\examples{
av_state <- load_file("../data/input/pp5 nieuw compleet.sav",log_level=3)
av_state <- var_main(av_state,c('SomBewegUur','SomPHQ'),criterion='BIC',log_level=3)
# av_state is the result of a call to var_main
contemporaneous_correlations_plot(av_state)
}

