% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_network}
\alias{generate_network}
\title{Return a JSON array of network data of a fitting model}
\usage{
generate_network(data, timestamp)
}
\arguments{
\item{data}{a data frame of 17 columns (ontspanning, opgewektheid, hier_en_nu, concentratie, beweging, iets_betekenen, humor, buiten_zijn, eigenwaarde, levenslust, onrust, somberheid, lichamelijk_ongemak, tekortschieten, piekeren, eenzaamheid, uw_eigen_factor) and 90 rows}

\item{timestamp}{the date of the first measurement in the format \code{'yyyy-mm-dd'}}
}
\value{
This function returns a string representing a json array of two networks.
}
\description{
This function uses repeated calls to \code{\link{var_main}} to find a fitting model for the data provided and then calls \code{\link{convert_to_graph}} to return a JSON representation of best valid model found.
}

