% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate_networks}
\alias{generate_networks}
\title{Return a JSON array of network data of a fitting model}
\usage{
generate_networks(data, timestamp, always_include = NULL, pairs = NULL,
  positive_variables = NULL, negative_variables = NULL, labels = list(),
  measurements_per_day = 3, max_network_size = 6)
}
\arguments{
\item{data}{a data frame. Each row is a measurement. Each column is an endogenous variable.}

\item{timestamp}{the date of the first measurement in the format \code{'yyyy-mm-dd'}. If multiple measurements are taken per day, the first measurement (row) in the data frame should correspond to the first measurement on that day.}

\item{always_include}{a vector of variable names that should always be included in the network if possible. Can also be \code{NULL}, in which case no variables are included by default.}

\item{pairs}{a vector of variable names in the form \code{c('pair1a','pair1b','pair2a','pair2b','pair3a',...)}. In other words: it is a vector of even length in which each two subsequent positions are seen as a pair. The pairs are treated such that only one pair or one variable of the pairs is included in the networks.}

\item{positive_variables}{a vector of names of variables that measure a positive effect (e.g., happiness). Variable names not occurring in the \code{positive_variables} or \code{negative_variables} list are considered neutral.}

\item{negative_variables}{a vector of names of variables that measure a negative effect (e.g., sadness). Variable names not occurring in the \code{positive_variables} or \code{negative_variables} list are considered neutral.}

\item{labels}{a list where keys are variable names and values are labels.}

\item{measurements_per_day}{an integer in [1,16] denoting the number of measurements per day. Defaults to 3.}

\item{max_network_size}{an integer in [2,6] denoting the number of nodes to include in the networks initially. Defaults to 6.}
}
\value{
This function returns a string representing a json array of two networks and an array of the top links.
}
\description{
This function uses repeated calls to \code{\link{var_main}} to find a fitting model for the data provided and then calls \code{\link{convert_to_graph}} to return a JSON representation of best valid model found.
}
\examples{
GN_COLUMNS <- c('ontspanning', 'opgewektheid', 'hier_en_nu', 'concentratie',
                'beweging', 'iets_betekenen', 'humor', 'buiten_zijn',
                'eigenwaarde', 'levenslust', 'onrust', 'somberheid',
                'lichamelijk_ongemak', 'tekortschieten', 'piekeren', 'eenzaamheid',
                'uw_eigen_factor')
d<-load_file("../data/input/DataDndN_nonimputed_voorAndo.sav",log_level=3)
d<-d$raw_data[,GN_COLUMNS]
timestamp <- '2014-03-01'
cat(generate_networks(data = d,
                  timestamp = timestamp,
                  always_include = 'uw_eigen_factor',
                  pairs = c('opgewektheid','onrust',
                            'somberheid','ontspanning',
                            'somberheid','onrust'),
                  positive_variables = c('opgewektheid','ontspanning','hier_en_nu',
                                         'concentratie', 'beweging','iets_betekenen',
                                         'humor', 'buiten_zijn','eigenwaarde', 'levenslust'),
                  negative_variables = c('onrust','somberheid','lichamelijk_ongemak',
                                         'tekortschieten','piekeren','eenzaamheid'),
                  labels = list(ontspanning = "Ontspanning",
                                opgewektheid = "Opgewektheid",
                                hier_en_nu = "In het hier en nu leven",
                                concentratie = "Concentratie",
                                beweging = "Beweging",
                                iets_betekenen = "Iets betekenen",
                                humor = "Humor",
                                buiten_zijn = "Buiten zijn",
                                eigenwaarde = "Eigenwaarde",
                                levenslust = "Levenslust",
                                onrust = "Onrust",
                                somberheid = "Somberheid",
                                lichamelijk_ongemak = "Lichamelijk ongemak",
                                tekortschieten = "Tekortschieten",
                                piekeren = "Piekeren",
                                eenzaamheid = "Eenzaamheid",
                                uw_eigen_factor = "Mijn eigen factor"),
                  measurements_per_day = 3,
                  max_network_size = 6))
}

