% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vargranger.r
\name{vargranger_plot}
\alias{vargranger_plot}
\title{Plot the Granger causality summary}
\usage{
vargranger_plot(av_state)
}
\arguments{
\item{av_state}{an object of class \code{av_state} that was the result of a call to \code{\link{var_main}}}
}
\description{
This function plots a summary of the Granger causality relations found in the valid models found. It is called as part of \code{\link{var_summary}}.
}
\examples{
av_state <- load_file("../data/input/pp5 nieuw compleet.sav",log_level=3)
av_state <- var_main(av_state,c('SomBewegUur','SomPHQ'),criterion='BIC',log_level=3)
# av_state is the result of a call to var_main
vargranger_plot(av_state)
}

