% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.r
\name{visualize_residuals}
\alias{visualize_residuals}
\title{Visualize the residuals of a VAR model}
\usage{
visualize_residuals(varest)
}
\arguments{
\item{varest}{the varest object.}
}
\description{
This function takes a varest object and plots the residuals and the squared residuals.
}
\examples{
d<-load_file("../data/input/pp4 nieuw compleet met 140min.sav",log_level=3)
d<-set_timestamps(d,date_of_first_measurement='2010-04-14')
d<-var_main(d,c('SomBewegUur','SomPHQ'),log_level=3,exogenous_variables='Work')
visualize_residuals(d$accepted_models[[1]]$varest)
}

