% Generated by roxygen2 (4.0.2): do not edit by hand
\name{impute_dataframe}
\alias{impute_dataframe}
\title{Impute missing values in a data.frame using EM imputation}
\usage{
impute_dataframe(df, net_cfg)
}
\arguments{
\item{df}{a data frame}

\item{net_cfg}{a net_cfg object providing metadata about the networks}
}
\value{
This function returns the modified data frame.
}
\description{
This function uses the Amelia package to impute a given data frame and return the imputed values.
}
\examples{
data <- generate_numerical_test_data(40)
data
impute_dataframe(data,list(measurements_per_day=1))
}

