% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/impute_dataframe.r
\name{impute_dataframe}
\alias{impute_dataframe}
\title{Impute missing values in a data.frame using EM imputation}
\usage{
impute_dataframe(df, measurements_per_day, repetitions = 30)
}
\arguments{
\item{df}{a data frame}

\item{measurements_per_day}{The number of measurements per day in the diary study}

\item{repetitions}{The amount of times the Amelia call should be averaged over. Defaults to 30. The actual number of imputations is five times the value for \code{repetitions}, since Amelia's values are already averaged over five runs.}
}
\value{
This function returns the modified data frame.
}
\description{
This function uses the Amelia package to impute a given data frame and return the imputed values.
}
\examples{
data <- generate_numerical_test_data(40)
data
impute_dataframe(data,measurements_per_day=1)
}

