% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/select_relevant_rows.r
\name{select_relevant_rows}
\alias{select_relevant_rows}
\title{Select and return the relevant rows}
\usage{
select_relevant_rows(data, timestamp, net_cfg)
}
\arguments{
\item{data}{a data frame}

\item{timestamp}{the date of the first measurement in the format \code{'yyyy-mm-dd'}}

\item{net_cfg}{a net_cfg object providing metadata about the networks}
}
\value{
This function returns a list with three elements, \code{data} is the modified data frame, \code{first_measurement_index} is the possibly adjusted first_measurement index, and \code{timestamp} is the possibly updated timestamp (day).
}
\description{
This function returns the data frame with the beginning and trailing rows that are all NA removed, it also returns updated \code{first_measurement_index} and \code{timestamp} variables. The timestamp may advance as a result of the data frame starting with three or more NA's.
}
\examples{
data <- data.frame(id=c(NA,NA,NA,1,1,NA),tijdstip=c(NA,NA,NA,3,4,NA),something=c(NA,NA,NA,9,16,NA))
data
net_cfg <- new_net_cfg()
net_cfg$measurements_per_day <- 2
select_relevant_rows(data,'2014-05-06',net_cfg)
}

