% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_timestamps.r
\name{set_timestamps}
\alias{set_timestamps}
\title{Add dummy variables for weekdays and day parts}
\usage{
set_timestamps(av_state, subset_id = 1, date_of_first_measurement,
  measurements_per_day = 1, log_level = 0,
  first_measurement_index = 1, add_days_as_exogenous = TRUE,
  add_dayparts_as_exogenous = TRUE, add_weekend_as_exogenous = FALSE)
}
\arguments{
\item{av_state}{an object of class \code{av_state}}

\item{subset_id}{either an integer subset index or the the value for the \code{id_field} column that was used in the \code{\link{group_by}} function. The \code{subset_id} argument is required if the data set is grouped into multiple data sets (i.e., if the \code{\link{group_by}} function was used), in which case the function works on the specified data subset.}

\item{date_of_first_measurement}{the date of the first measurement. This argument should be given in the format: \code{"yyyy-mm-dd"}, e.g., \code{"2004-03-28"}.}

\item{measurements_per_day}{how many measurements were taken per day. This default is 1. It is assumed that every day has exactly this amount of measurements, and that the first measurement in the dataset was the first measurement on that day.}

\item{log_level}{sets the minimum level of output that should be shown (a number between 0 and 3). A lower level means more verbosity. Specify a log_level of 3 to hide messages about the exogenous columns being added.}

\item{first_measurement_index}{is used to specify that the first day of measurements has fewer than \code{measurements_per_day} measurements. Here, we assume that the measurements in the data set still form a connected sequence. In other words, the assumption is that the missing measurements of the first day precede the first measurement in the data set. For example, by specifying \code{measurements_per_day = 3, first_measurement_index = 2}, the first measurement in the data set will be treated as the second measurement of that day. So the first two measurements in the data set will be tagged with \code{Afternoon} and \code{Evening}, and the third measurement in the data set will be tagged with \code{Morning} of the next day.}

\item{add_days_as_exogenous}{adds days as exogenous dummy variables to VAR models.}

\item{add_dayparts_as_exogenous}{adds day parts as exogenous dummy variables to VAR models.}

\item{add_weekend_as_exogenous}{adds one exogenous variable named \code{Weekend} to the VAR models. This variable is 1 for weekend days (Saturday and Sunday) and 0 otherwise. By specifying \code{add_days_as_exogenous = FALSE} and \code{add_weekend_as_exogenous = TRUE}, the weekend is used instead of day dummies in the evaluation of models.}
}
\value{
This function returns the modified \code{av_state} object.
}
\description{
This function adds dummy columns for weekdays (named \code{Sunday}, \code{Monday}, \code{Tuesday}, \code{Wednesday}, \code{Thursday}, \code{Friday} and \code{Saturday}) and day parts (\code{morning}, \code{afternoon}) to the given subset of the specified data set. These are used by \code{\link{var_main}} to find better models by removing cyclicity from the data set.
}
\examples{
\dontrun{
av_state <- load_file("../data/input/pp4 nieuw compleet met 140min.sav",log_level=3)
av_state <- set_timestamps(av_state,date_of_first_measurement="2010-04-14")

# an example with multiple measurements per day:
av_state <- load_file("../data/input/ID68 basisbestand.sav",log_level=3)
av_state <- set_timestamps(av_state,date_of_first_measurement="2012-07-12",
                           measurements_per_day=3)

# same data set, but one extra day because the first day only has one measurement
av_state <- load_file("../data/input/ID68 basisbestand.sav",log_level=3)
av_state <- set_timestamps(av_state,date_of_first_measurement="2012-07-12",
                           measurements_per_day=3,first_measurement_index=3)
}
}
