% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_file.r
\name{store_file}
\alias{store_file}
\title{Export a modified data set as an SPSS readable .sas file}
\usage{
store_file(av_state, filename, inline_data, file_type = c("SPSS", "STATA"))
}
\arguments{
\item{av_state}{an object of class \code{av_state}}

\item{filename}{the name that the exported files should be saved under. Do not include a path. When this argument is missing, the filename of the input file (used in \code{\link{load_file}}) is substituted.}

\item{inline_data}{boolean argument to determine whether data should be stored inline in the .sas script. This argument is optional, and defaults to \code{TRUE} if the data set has less than 80 columns, and to \code{FALSE} otherwise.}

\item{file_type}{sets the type of file export that should be used. Currently, only \code{'SPSS'} is supported.}
}
\description{
This function exports all subsets in the specified \code{av_state} to individual output files. All output files are subsequently packed in a .tar file. Data sets with less than 80 columns are stored inline in the .sas script, otherwise, the data is stored in separate files.
}
\examples{
\dontrun{
av_state <- load_file("../data/input/ID68 basisbestand.sav",log_level=3)
av_state <- add_trend(av_state)
av_state <- set_timestamps(av_state,date_of_first_measurement="2012-07-12",
                           measurements_per_day=3,log_level=3)
store_file(av_state)
}
}

